One more attempt to fix the logout.

+++ src/app/shared/services/auth.service.ts
@@ -21,8 +21,9 @@ export class AuthService {
     }
 
     logout() {
-        Cookie.deleteAll();
-
+        if (!environment.production) {
+            Cookie.deleteAll();
+        }
         window.location.href = environment.logoutUrl;
     }
